refactor: Simplify ManageAdminsUsersPage component and enhance user management features
- Removed redundant state variables and handlers related to user management, streamlining the component.
- Introduced a feature card system for better organization of admin functionalities, including adding, removing, and viewing users.
- Enhanced the user interface for managing admins and users, improving accessibility and usability.
- Updated the component structure to utilize a more modular approach, making it easier to maintain and extend in the future.
