+++ frontend/pages/manageusersandadmins/RemoveAdmin.js
@@ -1,0 +1,58 @@
+import React, { useState } from "react";
+import { BASE_API_URL } from "../../utils/apiurl";
+import { getToken } from "../../utils/auth";
+
+export default function RemoveAdmin({ userEmail, isSuperAdmin }) {
+  const [removeEmail, setRemoveEmail] = useState("");
+  const [removeStatus, setRemoveStatus] = useState("");
+
+  const handleRemoveAdmin = async (e) => {
+    e.preventDefault();
+    setRemoveStatus("Removing...");
+    try {
+      const res = await fetch(`${BASE_API_URL}/removeadmin`, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${getToken()}`,
+        },
+        body: JSON.stringify({
+          email: removeEmail,
+          requesterEmail: userEmail,
+        }),
+      });
+      if (res.ok) {
+        setRemoveStatus("Admin removed!");
+        setRemoveEmail("");
+      } else {
+        const data = await res.json();
+        setRemoveStatus(data.message || "Failed to remove admin");
+      }
+    } catch {
+      setRemoveStatus("Failed to remove admin");
+    }
+  };
+
+  if (!isSuperAdmin) return null;
+  return (
+    <div style={{ background: "#fff", borderRadius: 16, boxShadow: "0 2px 8px rgba(30,60,114,0.08)", padding: 32, marginBottom: 32 }}>
+      <h3 style={{ fontWeight: 700, fontSize: 22, marginBottom: 18, color: "#c0392b" }}>Remove Admin</h3>
+      <form onSubmit={handleRemoveAdmin}>
+        <div style={{ marginBottom: 12 }}>
+          <label>Email:</label><br />
+          <input
+            type="email"
+            required
+            value={removeEmail}
+            onChange={e => setRemoveEmail(e.target.value)}
+            style={{ width: "100%", padding: 8, borderRadius: 4, border: "1px solid #bbb" }}
+          />
+        </div>
+        <button type="submit" style={{ background: "#c0392b", color: "#fff", border: "none", borderRadius: 6, padding: "8px 24px", fontWeight: 600, cursor: "pointer" }}>
+          Remove
+        </button>
+        <div style={{ marginTop: 10, color: "#c0392b", fontWeight: 500 }}>{removeStatus}</div>
+      </form>
+    </div>
+  );
+}
